editCourseDetails now implements AdapterView.OnItemSelectedListener. This is because this class now contains an adapter set to a spinner that makes it so the user can simply select their courses category from a dropdown. Also had to override the two methods from the OnItemSelectedListener interface in order to implement it.
I've added functionality to the Add button which now allows the user to add more courses to the list. Still buggy (the buttons don't work) and I had to create a new templayout activity to use in order for the inflater to work with the view and acheive the desired result. I spent a lot of time trying to make it work by extracting already existing elements of our view but I haven't been able to find a solution where all the buttons aren't imported so I just created a child layout called activity_templayout.xml

Rearranged the buttons on the main screen so that they are now in a vertical list

Removed the other linearlayouts from the activity_templates_screen as they are no longer needed